Added return values to transmit functions
Changed  writePot0, writePot1 and writePots methods so that they return
the uint8_t status code from the Wire library.   Changed configure so it
returns a boolean succes or failure value.

Also moved GetSlaveAddress to a public method.  It doesn't hurt anyone
or anything and I needed it there for a project.