Added support for constructing by address pin settings
Added GetSlaveAddress() override and constructors to MaximDS::Controller
that allow the user to construct a Controller object with the I2C address generated
based on the supplied address pin states.

Also getI2CBusAddress() function to get the I2C bus address assigned to the device for
testing and debugging purposes.

Also corrected some spelling mistakes in the documentation and fixed some formatting